pyproject.toml
[project]
name = "dialectica-agents"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.175.0",
    "google-generativeai>=0.8.5",
    "groq>=0.31.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
